// frontend/src/client/types.gen.ts
export type AppointmentCreate = {
    patient_name: string;
    patient_id_number: string;
    patient_phone: string;
    patient_email?: (string | null);
    appointment_time: string;
    status?: string;
    hospital_id: string;
    doctor_id: string;
};
// ...
export type UserValidation = {
    name: string;
    id_number: string;
    phone: string;
    email?: (string | null);
};